+++ src/akkasls.ts
@@ -1,129 +1,68 @@
-/**
- * Auth CLI actions
- */
-import { login } from './akkasls/commands/auth/login';
-import { logout } from './akkasls/commands/auth/logout';
-import { currentLogin } from './akkasls/commands/auth/currentlogin';
-import { listAuthTokens } from './akkasls/commands/auth/tokens/list';
-import { revokeAuthToken } from './akkasls/commands/auth/tokens/revoke';
-
-/**
- * Docker CLI actions
- */
-import { addDockerCredentials } from './akkasls/commands/docker/addcredentials';
-import { deleteDockerCredentials } from './akkasls/commands/docker/deletecredentials';
-import { listDockerCredentials } from './akkasls/commands/docker/listcredentials';
-
-/**
- * Projects CLI actions
- */
-import { listProjects } from './akkasls/commands/projects/list';
-import { newProject } from './akkasls/commands/projects/new';
-
-/**
- * Roles CLI actions
- */
-import { deleteInvite } from './akkasls/commands/roles/invitations/delete';
-import { addInvite } from './akkasls/commands/roles/invitations/inviteuser';
-import { listInvites } from './akkasls/commands/roles/invitations/list';
-import { listMembers } from './akkasls/commands/roles/listbindings';
-
-/**
- * Services CLI actions
- */
-import { deployService } from './akkasls/commands/services/deploy';
-import { exposeService } from './akkasls/commands/services/expose';
-import { listServices } from './akkasls/commands/services/list';
-import { undeployService } from './akkasls/commands/services/undeploy';
-import { unexposeService } from './akkasls/commands/services/unexpose';
-
-/**
- * Local Proxy CLI extensions
- */
-import { startLocalProxy } from './akkasls/extensions/commands/localproxy/start';
-import { stopLocalProxy } from './akkasls/extensions/commands/localproxy/stop';
-
-/**
- * Utils
- */
-import { Shell } from './utils/shell/shell';
-import { window } from 'vscode';
-
-/**
- * Wizards
- */
-import { projectPicker } from './components/wizards/projectPicker';
-import { inputBox } from './components/wizards/inputBox';
-import { dockerCredentialsPicker } from './components/wizards/dockerCredentialsPicker';
-import { servicePicker } from './components/wizards/servicePicker';
-import { invitePicker } from './components/wizards/invitePicker';
-import { tokenPicker } from './components/wizards/tokenPicker';
-
-/**
- * DataTypes
- */
-import { CurrentLogin } from './akkasls/datatypes/auth/currentlogin';
-import { Token } from './akkasls/datatypes/auth/tokens';
-import { Credential } from './akkasls/datatypes/docker/credentials';
-import { ShellResult } from './utils/shell/datatypes';
-import { Project } from './akkasls/datatypes/projects/project';
-import { Invite } from './akkasls/datatypes/roles/invitations/invite';
-import { Member } from './akkasls/datatypes/roles/member';
-import { Service } from './akkasls/datatypes/services/service';
-
-/**
- * Explorer viewa
- */
-import { ProjectExplorer } from './components/views/projectexplorer/explorer';
-import { ConfigExplorer } from './components/views/configexplorer/explorer';
-
-export const CONSOLE_URL = "https://console.cloudstate.com/project";
+import { login, logout, currentLogin, listAuthTokens, revokeAuthToken, addDockerCredentials, deleteDockerCredentials, listDockerCredentials, listProjects, newProject, deleteInvite, addInvite, listInvites, listMembers, deployService, exposeService, listServices, undeployService, unexposeService, CurrentLogin, Token, Credential, Project, Invite, Member, Service, ShellResult } from '@retgits/akkasls-nodewrapper';
+import { ASLocal } from './plugins/local/plugin';
+import { DOCKER_ITEM_TYPE } from './plugins/projectexplorer/dockerCredentialItem';
+import { SERVICE_ITEM_TYPE } from './plugins/projectexplorer/serviceItem';
+import { window, workspace } from 'vscode';
+import { projectPicker } from './plugins/wizards/projectPicker';
+import { inputBox } from './plugins/wizards/inputBox';
+import { dockerCredentialsPicker } from './plugins/wizards/dockerCredentialsPicker';
+import { servicePicker } from './plugins/wizards/servicePicker';
+import { invitePicker } from './plugins/wizards/invitePicker';
+import { tokenPicker } from './plugins/wizards/tokenPicker';
+import { ProjectExplorer } from './plugins/projectexplorer/projectExplorer';
+import { ConfigExplorer } from './plugins/configexplorer/configExplorer';
+import { logger } from './utils/logger';
+import { functionPicker } from './plugins/wizards/functionPicker';
 
 export class AkkaServerless {
-    private _shell: Shell;
     private _projectExplorer: ProjectExplorer;
     private _configExplorer: ConfigExplorer;
 
-    /**
-     * Create a new instance of AkkaServerless which handles communication between the VS code extension and the underlying commands
-     * @param shell a shell that handles execution of commands
-     */
-    constructor(shell: Shell) {
-        this._shell = shell;
+    // TODO: Add the CommandInput as exported type on akkasls-nodewrapper
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    private _getCommandInput(): any {
+        return {
+            dryrun: workspace.getConfiguration('akkaserverless').get<boolean>('dryrun'),
+            silent: workspace.getConfiguration('akkaserverless').get<boolean>('logOutput'),
+            configFile: workspace.getConfiguration('akkaserverless').get<string>('configFile'),
+            context: workspace.getConfiguration('akkaserverless').get<string>('context')
+        };
     }
 
-    registerProjectExplorer(pe: ProjectExplorer) {
+    registerProjectExplorer(pe: ProjectExplorer): void {
         this._projectExplorer = pe;
     }
 
-    registerConfigExplorer(ce: ConfigExplorer) {
+    registerConfigExplorer(ce: ConfigExplorer): void {
         this._configExplorer = ce;
     }
 
-    async login() {
-        let result = await login(this._shell);
+    async login(): Promise<void> {
+        const result = await login(this._getCommandInput());
         if (result.code !== 0) {
             window.showErrorMessage(result.stderr);
         }
     }
 
-    async logout() {
-        let result = await logout(this._shell);
+    async logout(): Promise<void> {
+        const result = await logout(this._getCommandInput());
         if (result.code !== 0) {
             window.showErrorMessage(result.stderr);
         }
     }
 
     async getCurrentLogin(): Promise<CurrentLogin> {
-        return currentLogin(this._shell);
+        const res = await currentLogin(this._getCommandInput());
+        return res.response as CurrentLogin;
     }
 
     async listAuthTokens(): Promise<Token[]> {
-        return listAuthTokens(this._shell);
+        const res = await listAuthTokens(this._getCommandInput());
+        return res.response as Token[];
     }
 
     async revokeAuthToken(tokenID: string): Promise<ShellResult> {
-        let result = await revokeAuthToken(tokenID, this._shell);
+        const result = await revokeAuthToken(tokenID, this._getCommandInput());
         if (result.code !== 0) {
             window.showErrorMessage(result.stderr);
         }
@@ -132,7 +71,7 @@ export class AkkaServerless {
     }
 
     async addDockerCredentials(projectID: string, credentials: string): Promise<ShellResult> {
-        let result = await addDockerCredentials(projectID, credentials, this._shell);
+        const result = await addDockerCredentials(projectID, credentials, this._getCommandInput());
         if (result.code !== 0) {
             window.showErrorMessage(result.stderr);
         }
@@ -140,22 +79,25 @@ export class AkkaServerless {
         return result;
     }
 
-    async deleteDockerCredentials(projectID: string, credentialID: string): Promise<ShellResult> {
-        let result = await deleteDockerCredentials(projectID, credentialID, this._shell);
+    async deleteDockerCredentials(projectID: string, credentialID: string): Promise<void> {
+        if (credentialID === DOCKER_ITEM_TYPE) {
+            return;
+        }
+
+        const result = await deleteDockerCredentials(projectID, credentialID, this._getCommandInput());
         if (result.code !== 0) {
             window.showErrorMessage(result.stderr);
         }
         this._projectExplorer.refresh();
-        return result;
     }
 
     async listDockerCredentials(projectID: string): Promise<Credential[]> {
-        let result = await listDockerCredentials(projectID, this._shell);
-        return result;
+        const res = await listDockerCredentials(projectID, this._getCommandInput());
+        return res.response as Credential[];
     }
 
     async createNewProject(name: string, description: string): Promise<ShellResult> {
-        let result = await newProject(name, description, this._shell);
+        const result = await newProject(name, description, this._getCommandInput());
         if (result.code !== 0) {
             window.showErrorMessage(result.stderr);
         }
@@ -164,12 +106,12 @@ export class AkkaServerless {
     }
 
     async listProjects(): Promise<Project[]> {
-        let result = await listProjects(this._shell);
-        return result;
+        const res = await listProjects(this._getCommandInput());
+        return res.response as Project[];
     }
 
     async deleteInvite(projectID: string, emailAddress: string): Promise<ShellResult> {
-        let result = await deleteInvite(projectID, emailAddress, this._shell);
+        const result = await deleteInvite(projectID, emailAddress, this._getCommandInput());
         if (result.code !== 0) {
             window.showErrorMessage(result.stderr);
         }
@@ -177,8 +119,8 @@ export class AkkaServerless {
         return result;
     }
 
-    async addInvite(projectID: string, emailAddress: string) : Promise<ShellResult> {
-        let result = await addInvite(projectID, emailAddress, this._shell);
+    async addInvite(projectID: string, emailAddress: string): Promise<ShellResult> {
+        const result = await addInvite(projectID, emailAddress, this._getCommandInput());
         if (result.code !== 0) {
             window.showErrorMessage(result.stderr);
         }
@@ -187,17 +129,17 @@ export class AkkaServerless {
     }
 
     async listInvites(projectID: string): Promise<Invite[]> {
-        let result = await listInvites(projectID, this._shell);
-        return result;
+        const res = await listInvites(projectID, this._getCommandInput());
+        return res.response as Invite[];
     }
 
     async listMembers(projectID: string): Promise<Member[]> {
-        let result = await listMembers(projectID, this._shell);
-        return result;
+        const res = await listMembers(projectID, this._getCommandInput());
+        return res.response as Member[];
     }
 
-    async deployService(service:string, image:string, projectID:string): Promise<ShellResult> {
-        let result = await deployService(service, image, projectID, this._shell);
+    async deployService(service: string, image: string, projectID: string): Promise<ShellResult> {
+        const result = await deployService(service, image, projectID, this._getCommandInput());
         if (result.code !== 0) {
             window.showErrorMessage(result.stderr);
         }
@@ -205,8 +147,8 @@ export class AkkaServerless {
         return result;
     }
 
-    async exposeService(service:string, flags:string, projectID:string): Promise<ShellResult> {
-        let result = await exposeService(service, flags, projectID, this._shell);
+    async exposeService(service: string, flags: string, projectID: string): Promise<ShellResult> {
+        const result = await exposeService(service, flags, projectID, this._getCommandInput());
         if (result.code !== 0) {
             window.showErrorMessage(result.stderr);
         }
@@ -214,12 +156,12 @@ export class AkkaServerless {
     }
 
     async listServices(projectID: string): Promise<Service[]> {
-        let result = await listServices(projectID, this._shell);
-        return result;
+        const res = await listServices(projectID, this._getCommandInput());
+        return res.response as Service[];
     }
 
-    async undeployService(service:string, projectID:string): Promise<ShellResult> {
-        let result = await undeployService(service, projectID, this._shell);
+    async undeployService(service: string, projectID: string): Promise<ShellResult> {
+        const result = await undeployService(service, projectID, this._getCommandInput());
         if (result.code !== 0) {
             window.showErrorMessage(result.stderr);
         }
@@ -227,132 +169,201 @@ export class AkkaServerless {
         return result;
     }
 
-    async unexposeService(service:string, hostname:string, projectID:string): Promise<ShellResult> {
-        let result = await unexposeService(service, hostname, projectID, this._shell);
+    async unexposeService(service: string, hostname: string, projectID: string): Promise<ShellResult> {
+        const result = await unexposeService(service, hostname, projectID, this._getCommandInput());
         if (result.code !== 0) {
             window.showErrorMessage(result.stderr);
         }
         return result;
     }
 
-    async startLocalProxy(configfile: string, dockerImageUser: string): Promise<ShellResult> {
-        let result = await startLocalProxy(configfile, dockerImageUser, this._shell);
-        if (result.code !== 0) {
-            window.showErrorMessage(result.stderr);
+    async startLocalProxy(configfile: string): Promise<void> {
+        const asLocal = new ASLocal(configfile);
+        try {
+            const func = await functionPicker('pick a function to start locally', asLocal.getServices());
+            asLocal.startLocalProxy(func);
+        } catch (e) {
+            window.showErrorMessage(e);
+            logger.error(e);
         }
-        return result;
     }
 
-    async stopLocalProxy(configfile: string, dockerImageUser: string): Promise<ShellResult> {
-        let result = await stopLocalProxy(configfile, dockerImageUser, this._shell);
-        if (result.code !== 0) {
-            window.showErrorMessage(result.stderr);
+    async stopLocalProxy(configfile: string): Promise<void> {
+        const asLocal = new ASLocal(configfile);
+        try {
+            const func = await functionPicker('pick a function to stop', asLocal.getServices());
+            asLocal.stopLocalProxy(func);
+        } catch (e) {
+            window.showErrorMessage(e);
+            logger.error(e);
         }
-        return result;
     }
 
-    async addDockerCredentialsWizard(projectID?: string) {
-        if(!projectID) {
-            projectID = await projectPicker('pick a project to add credentials to...', await this.listProjects());
+    async addDockerCredentialsWizard(projectID?: string): Promise<void> {
+        if (!projectID) {
+            try {
+                projectID = await projectPicker('pick a project to add credentials to...', await this.listProjects());
+                const credentials = await inputBox('--docker-server <> --docker-username <>> --docker-password <>', 'enter your credential string...');
+                this.addDockerCredentials(projectID, credentials);
+            } catch (e) {
+                window.showErrorMessage(e);
+                logger.error(e);
+            }
         }
-
-        let credentials = await inputBox('--docker-server <> --docker-username <>> --docker-password <>', 'enter your credential string...');
-        this.addDockerCredentials(projectID, credentials);
     }
 
-    async deleteDockerCredentialsWizard(projectID?: string, credentialID?: string) {
-        if(!projectID) {
-            projectID = await projectPicker('pick a project to remove credentials from...', await this.listProjects());
-        }
+    async deleteDockerCredentialsWizard(projectID?: string, credentialID?: string): Promise<void> {
+        try {
+            if (!projectID) {
+                projectID = await projectPicker('pick a project to remove credentials from...', await this.listProjects());
+            }
 
-        if(!credentialID) {
-            credentialID = await dockerCredentialsPicker(await this.listDockerCredentials(projectID));
-        }
+            if (!credentialID) {
+                credentialID = await dockerCredentialsPicker(await this.listDockerCredentials(projectID));
+            }
 
-        this.deleteDockerCredentials(projectID, credentialID);
+            this.deleteDockerCredentials(projectID, credentialID);
+        } catch (e) {
+            window.showErrorMessage(e);
+            logger.error(e);
+            return;
+        }
     }
 
-    async deployServiceWizard(projectID?: string) {
-        if(!projectID) {
-            projectID = await projectPicker('pick a project to deploy your service to...', await this.listProjects());
+    async deployServiceWizard(projectID?: string): Promise<void> {
+        try {
+            if (!projectID) {
+                projectID = await projectPicker('pick a project to deploy your service to...', await this.listProjects());
+            }
+            const service = await inputBox('', 'type your service name...');
+            const image = await inputBox('', 'type your docker image url...');
+            this.deployService(service, image, projectID);
+        } catch (e) {
+            window.showErrorMessage(e);
+            logger.error(e);
+            return;
         }
-
-        let service = await inputBox('', 'type your service name...');
-        let image = await inputBox('', 'type your docker image url...');
-
-        this.deployService(service, image, projectID);
     }
 
-    async undeployServiceWizard(projectID?: string, service?: string) {
-        if(!projectID) {
-            projectID = await projectPicker('pick a project to undeploy your service from...', await this.listProjects());
+    async undeployServiceWizard(projectID?: string, service?: string): Promise<void> {
+        if (service === SERVICE_ITEM_TYPE) {
+            return;
         }
 
-        if(!service) {
-            service = await servicePicker('pick the service to undeploy...', await this.listServices(projectID));
-        }
+        try {
+            if (!projectID) {
+                projectID = await projectPicker('pick a project to undeploy your service from...', await this.listProjects());
+            }
 
-        this.undeployService(service, projectID);
+            if (!service) {
+                service = await servicePicker('pick the service to undeploy...', await this.listServices(projectID));
+
+            }
+            this.undeployService(service, projectID);
+        } catch (e) {
+            window.showErrorMessage(e);
+            logger.error(e);
+            return;
+        }
     }
 
-    async exposeServiceWizard(projectID?: string, service?:string) {
-        if(!projectID) {
-            projectID = await projectPicker('pick a project to expose your service from...', await this.listProjects());
+    async exposeServiceWizard(projectID?: string, service?: string): Promise<void> {
+        if (service === SERVICE_ITEM_TYPE) {
+            return;
         }
 
-        if(!service) {
-            service = await servicePicker('pick the service to expose...', await this.listServices(projectID));
+        try {
+            if (!projectID) {
+                projectID = await projectPicker('pick a project to expose your service from...', await this.listProjects());
+            }
+
+            if (!service) {
+                service = await servicePicker('pick the service to expose...', await this.listServices(projectID));
+            }
+
+            const flags = await inputBox('--enable-cors', 'type any additional flags you might want...');
+            this.exposeService(service, flags, projectID);
+        } catch (e) {
+            window.showErrorMessage(e);
+            logger.error(e);
+            return;
         }
-
-        let flags = await inputBox('--enable-cors', 'type any additional flags you might want...');
-
-        this.exposeService(service, flags, projectID);
     }
 
-    async unexposeServiceWizard(projectID?: string, service?:string) {
-        if(!projectID) {
-            projectID = await projectPicker('pick a project to unexpose your service from...', await this.listProjects());
+    async unexposeServiceWizard(projectID?: string, service?: string): Promise<void> {
+        if (service === SERVICE_ITEM_TYPE) {
+            return;
         }
 
-        if(!service) {
-            service = await servicePicker('pick the service to unexpose...', await this.listServices(projectID));
+        try {
+            if (!projectID) {
+                projectID = await projectPicker('pick a project to unexpose your service from...', await this.listProjects());
+            }
+
+            if (!service) {
+                service = await servicePicker('pick the service to unexpose...', await this.listServices(projectID));
+            }
+
+            const hostname = await inputBox('', 'type the hostname you want to remove...');
+            this.unexposeService(service, hostname, projectID);
+        } catch (e) {
+            window.showErrorMessage(e);
+            logger.error(e);
+            return;
         }
-
-        let hostname = await inputBox('', 'type the hostname you want to remove...');
-
-        this.unexposeService(service, hostname, projectID);
     }
 
-    async inviteUserWizard(projectID?: string) {
-        if(!projectID) {
-            projectID = await projectPicker('pick a project to invite a new person to...', await this.listProjects());
+    async inviteUserWizard(projectID?: string): Promise<void> {
+        try {
+            if (!projectID) {
+                projectID = await projectPicker('pick a project to invite a new person to...', await this.listProjects());
+            }
+
+            const emailAddress = await inputBox('', 'type the email address of the person you want to invite...');
+            this.addInvite(projectID, emailAddress);
+        } catch (e) {
+            window.showErrorMessage(e);
+            logger.error(e);
+            return;
         }
-
-        let emailAddress = await inputBox('', 'type the email address of the person you want to invite...');
-
-        this.addInvite(projectID, emailAddress);
     }
 
-    async deleteInviteWizard(projectID?: string, emailAddress?: string) {
-        if(!projectID) {
-            projectID = await projectPicker('pick a project to invite a new person to...', await this.listProjects());
-        }
+    async deleteInviteWizard(projectID?: string, emailAddress?: string): Promise<void> {
+        try {
+            if (!projectID) {
+                projectID = await projectPicker('pick a project to invite a new person to...', await this.listProjects());
+            }
 
-        if(!emailAddress) {
-            emailAddress = await invitePicker('pick the email address to uninvite...', await this.listInvites(projectID));
-        }
+            if (!emailAddress) {
+                emailAddress = await invitePicker('pick the email address to uninvite...', await this.listInvites(projectID));
+            }
 
-        this.deleteInvite(projectID, emailAddress);
+            this.deleteInvite(projectID, emailAddress);
+        } catch (e) {
+            window.showErrorMessage(e);
+            logger.error(e);
+            return;
+        }
     }
 
-    async newProjectWizard() {
-        let name = await inputBox('', 'type the name of the project you want to create...');
-        let description = await inputBox('', 'type the description of the project you want to create...');
-        this.createNewProject(name, description);
+    async newProjectWizard(): Promise<void> {
+        try {
+            const name = await inputBox('', 'type the name of the project you want to create...');
+            const description = await inputBox('', 'type the description of the project you want to create...');
+            this.createNewProject(name, description);
+        } catch (e) {
+            window.showErrorMessage(e);
+            logger.error(e);
+        }
     }
 
-    async revokeAuthTokenWizard() {
-        let tokenID = await tokenPicker('pick the token you want to revoke...', await this.listAuthTokens());
-        this.revokeAuthToken(tokenID);
+    async revokeAuthTokenWizard(): Promise<void> {
+        try {
+            const tokenID = await tokenPicker('pick the token you want to revoke...', await this.listAuthTokens());
+            this.revokeAuthToken(tokenID);
+        } catch (e) {
+            window.showErrorMessage(e);
+            logger.error(e);
+        }
     }
 }