+++ src/akkasls/extensions/datatypes/tool.ts
@@ -1,7 +1,0 @@
-export interface Tool {
-    name: string;
-    versionCmd: string;
-    currentVersion?: string;
-    infoURL?: string;
-    updateURL?: string;
-}